requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.
django-avatar==1.0.5
PIL==1.1.7
South==0.7.4
django-activity-stream==0.4.3
django-authority==0.4

# Informal requirements, the apps we copied the source code
# django-friends
